ROI Extraction/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
numpy>=1.23.0
pandas>=1.5.0
pillow>=9.0.0
matplotlib>=3.7.0
opencv-python>=4.7.0
scikit-learn>=1.2.0
pycocotools>=2.0.6
tqdm>=4.65.0
PyYAML>=6.0
albumentations>=1.3.0 